third_party/WebKit/LayoutTests/fast/events/keyboardevent-key.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// Note: This test is only testing synthetic DomKey for eventSender, for real
// DomKey generated on each platform please refer to individual tests.
// e.g. platform_key_map_win_unittest.cc, web_input_event_builders_mac_unittest.mm
test(function() {
    if (!window.eventSender) {
        document.write('This test requires eventSender');
        return;
    }

    var lastKeyDown, lastKeyPress, lastKeyUp;
    function recordKeyEvent(event) {
        if (event.type === "keydown")
            lastKeyDown = event;
        else if (event.type === "keypress")
            lastKeyPress = event;
        else if (event.type === "keyup")
            lastKeyUp = event;
    }
    document.addEventListener('keydown', recordKeyEvent, false);
    document.addEventListener('keypress', recordKeyEvent, false);
    document.addEventListener('keyup', recordKeyEvent, false);


    function testDomKey(keyString, expectedDomKey, shouldTestKeyPress, modifiers) {
        lastKeyDown = null;
        lastKeyPress = null;
        lastKeyUp = null;
        eventSender.keyDown(keyString, modifiers);
        assert_equals(lastKeyDown.key, expectedDomKey);
        assert_equals(lastKeyUp.key, expectedDomKey);
        if (shouldTestKeyPress)
            assert_equals(lastKeyPress.key, expectedDomKey);
    }
    // Alpha&Numeric Keys
    testDomKey('A', 'A', true, ['shiftKey']);
    testDomKey('g', 'g', true);
    testDomKey('h', 'h', true);
    testDomKey('1', '1', true);
    testDomKey('(', '(', true, ['shiftKey']);

    // Modifier Keys
    testDomKey('leftControl', 'Control', false);
    testDomKey('rightControl', 'Control', false);
    testDomKey('leftShift', 'Shift', false);
    testDomKey('rightShift', 'Shift', false);
    testDomKey('leftAlt', 'Alt', false);
    testDomKey('rightAlt', 'Alt', false);
    testDomKey('numLock', 'NumLock', false);

    // Whitespace Keys
    testDomKey('\n', 'Enter', true);
    testDomKey('\t', 'Tab', false);
    
    // Navigation keys
    testDomKey('rightArrow', 'ArrowRight', false);
    testDomKey('downArrow', 'ArrowDown', false);
    testDomKey('leftArrow', 'ArrowLeft', false);
    testDomKey('upArrow', 'ArrowUp', false);
    testDomKey('home', 'Home', false);
    testDomKey('end', 'End', false);
    testDomKey('pageUp', 'PageUp', false);
    testDomKey('pageDown', 'PageDown', false);
    
    // Editing Keys
    testDomKey('backspace', 'Backspace', false);
    testDomKey('insert', 'Insert', false);
    testDomKey('delete', 'Delete', false);
      
    // Device Keys
    testDomKey('printScreen', 'PrintScreen', false);
    
    // General Purpose Function Section
    testDomKey('F1', 'F1', false);
    testDomKey('F6', 'F6', false);
    testDomKey('F12', 'F12', false);
    
    // UI Keys
    testDomKey('escape', 'Escape', false);
    testDomKey('menu', 'ContextMenu', false);
}, 'Testing synthetic DomKey for eventSender.');
</script>
